Change test in git email check that is run in prompt to prevent glitch
The previous implementation used [[ -d .git ]] && .... as the test.
When an invalid test like '[[ "3" >= "4" ]]' (the invalid test is the >=
- could also be <= or => or any other typo, I think
Here is an example of that error:

  dotan@platelet infra-platform-v2 ~/MyCode/bits-and-bobs main dotan@corky.net>[[ "3" >= "4" ]]
  bash: syntax error in conditional expression
  bash: command substitution: line 3: syntax error near unexpected token `-d'
  bash: command substitution: line 3: `[[ -d .git ]] && git config user.email)'
  bash: command substitution: line 2: syntax error near unexpected token `-d'
  bash: command substitution: line 2: `[[ -d .git ]] && git config user.email'
  dotan@platelet infra-platform-v2 ~/MyCode/bits-and-bobs main >]]
  bash: syntax error near unexpected token `]]'
  dotan@platelet infra-platform-v2 ~/MyCode/bits-and-bobs main dotan@corky.net>

The glitch can be stopped by typing some characters (`]]` above works),
but may persist if I just hit ENTER or type some other commands (like
`echo hello`, for example).

Not sure why the code I type breaks the command in the prompt, but I
replaced `[[ -d .git ]] && ...` with `test -d .git && ...` and the
glitch stopped recurring.